stash: introduce new abstraction for persistent metadata storage
The new `stash` crate provides a durable metadata store called a
`Stash`. A stash is like a differential arrangement but its state is
durable.

The idea is that each layer of the platform will store its durable state
in a `Stash`. The API of `Stash` is intentionally very similar to the
intended API of a STORAGE collection. Once STORAGE is stable enough to
serve as a source of truth, we intend to swap all stashes for STORAGE
collections.